0.0.4

+++ contracts/token/interfaces/IToken.sol
@@ -5,10 +5,10 @@ interface IToken {
     function totalSupply() external view returns(uint256);
     function balanceOf(address tokenOwner) external view returns (uint256);
     function allowance(address tokenOwner, address spender) external view returns (uint256);
-    function transfer(address to, uint tokens) external returns (bool);
-    function approve(address spender, uint tokens) external returns (bool);
-    function transferFrom(address from, address to, uint tokens) external returns (bool);
+    function transfer(address to, uint256 tokens) external returns (bool);
+    function approve(address spender, uint256 tokens) external returns (bool);
+    function transferFrom(address from, address to, uint256 tokens) external returns (bool);
 
-    event Transfer(address indexed from, address indexed to, uint tokens);
-    event Approval(address indexed tokenOwner, address indexed spender, uint tokens);
+    event Transfer(address indexed from, address indexed to, uint256 tokens);
+    event Approval(address indexed tokenOwner, address indexed spender, uint256 tokens);
 }
